requirements.txt
pyspark
psycopg2-binary
python-dotenv
Minio
pandas
pyarrow
clickhouse-connect
clickhouse-driver
apache-airflow
apache-airflow-providers-apache-spark
